refactor: move `updatePersonProperties`
